perf(tokenizer): remove byte-path uppercase overhead
